site-cookbooks iptables add

+++ site-cookbooks/wordpress/recipes/default.rb
@@ -7,8 +7,8 @@
 # All rights reserved - Do Not Redistribute
 #
 filename = "#{node['wordpress']['wp_tar_name']}"
-install_dir = "#{node['wordpress']['wp_dir']}"
-document_root = "/var/www/rfa/"
+install_dir = "#{node['wordpress']['install_dir']}"
+document_root = "#{node['wordpress']['document_root']}"
 remote_uri = "#{node['wordpress']['wp_tar_uri']}"
 file_checksum = "#{node['wordpress']['wp_tar_sum']}"
 httpd_conf_dir = "#{node['wordpress']['httpd_conf_dir']}"
@@ -23,19 +23,18 @@ htpasswd_pw = "#{node['wordpress']['htpasswd_pw']}"
 #  end
 #end
 
-#インストールディレクトリ作成
-directory "#{install_dir}" do
-  recursive true
-  mode 0755
-  owner "apache"
-  group "apache"
-  action :create
-end
-
 #WordPressダウンロード
 remote_file "/tmp/#{filename}" do
   source "#{remote_uri}"
   checksum "#{file_checksum}"
+end
+
+#ドキュメントルート設定
+directory "#{document_root}" do
+  recursive true
+  mode 0755
+  owner "apache"
+  group "apache"
 end
 
 #WordPress配置
@@ -44,10 +43,9 @@ script "install_wordpress" do
   user        "root"
   code <<-EOL
     install -d #{install_dir}
-    tar zxvf /tmp/#{filename} -C #{document_root}
+    tar zxvf /tmp/#{filename} -C #{install_dir}
   EOL
-#  not_if { File.exists?("") }
-#    tar zxvf /tmp/#{filename} -C #{install_dir}
+  not_if { File.exists?("#{document_root}") }
 end
 
 #httpd.conf配置
@@ -61,7 +59,7 @@ end
 service "httpd"
 
 #wordpressディレクトリpermission変更
-directory "#{install_dir}/wordpress" do
+directory "#{install_dir}" do
   recursive true
   mode 0755
   owner "apache"
